+++ frame/src/main/java/com/weicoder/frame/entity/EntityUserId.java
@@ -11,11 +11,11 @@ public interface EntityUserId {
 	 * 获得用户ID
 	 * @return 用户ID
 	 */
-	Integer getUserId();
+	Long getUserId();
 
 	/**
 	 * 设置用户ID
 	 * @param userId 用户ID
 	 */
-	void setUserId(Integer userId);
+	void setUserId(Long userId);
 }
